backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
aiofiles
google-cloud-sql-connector
mysql-connector-python==8.2.0
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.10.0
gunicorn

